Migrate snippet message to new structure
