--- app/message-snippet/server/index.js
@@ -1,3 +1,0 @@
-import './startup/settings';
-import './methods/snippetMessage';
-import './requests';
