+++ server/api/methods/index.js
@@ -5,3 +5,4 @@ import './2fa/regenerateCodes';
 import './2fa/validateTempToken';
 import './mentions/getUserMentionsByChannel';
 import './messages/unreadMessages';
+import './messages/snippetMessage';
