+++ server/api/methods/messages/snippetMessage.js
@@ -1,0 +1,7 @@
+import { Meteor } from 'meteor/meteor';
+
+import { snippetMessage } from '../../../services/messages/functions/snippetMessage';
+
+Meteor.methods({
+	snippetMessage,
+});
