+++ server/settings/settingfiles/index.ts
@@ -1,3 +1,4 @@
 import './2fa.settings';
 import './api.settings';
 import './pinmessage.settings';
+import './snippetmessage.settings';
