+++ server/settings/settingfiles/snippetmessage.settings.ts
@@ -1,7 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 
-import { settings } from '../../../settings';
-import { Permissions } from '../../../models';
+import { settings } from '..';
+import { Permissions } from '../../models';
 
 Meteor.startup(function() {
 	settings.add('Message_AllowSnippeting', false, {
